+++ src/models/task.ts
@@ -6,14 +6,18 @@ export default class task {
     project_id!: Number;
     section_id!: Number;
     content!: string;
+    description!: string;
     completed!: Boolean;
     label_ids!: Number[];
+    parent_id!: Number;
     parent!: Number;
     order!: Number;
     priority!: Number;
     due!: DueDate;
     url!: Url;
     comment_count!: Number;
+    assignee!: Number;
+    assigner!: Number;
 
     public static deserialize(json: any) : task {
         //TODO : Validate before returning